[视频] 修改视频主题验证

+++ app/Commands/Thread/CreateThread.php
@@ -88,7 +88,7 @@ class CreateThread
         Arr::set($this->data, 'attributes.content', $content);
 
         // 存在审核敏感词/发布视频主题时，将主题放入待审核
-        $thread->type = Arr::get($this->data, 'attributes.type', 0);
+        $thread->type = (int) Arr::get($this->data, 'attributes.type', 0);
         if ($censor->isMod || $thread->type == 2) {
             $thread->is_approved = 0;
         }
